fetch部分をカスタムフック化
