+++ src/pages/detail.tsx
@@ -1,27 +1,8 @@
-import { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
 import { TodoDetailTemplate } from "../components/templates/TodoDetailTemplate";
-import { SERVER_BASE_PATH } from "../constants/constants";
-import { TodoType } from "../types/TodoType";
+import { useFetchTodo } from "../hooks/useFetchTodo";
 
 export const TodoDetailPage = () => {
-  const { id } = useParams();
-  const [todo, setTodo] = useState<TodoType | null>(null);
-
-  const fetchTodo = async () => {
-    try {
-      const response = await fetch(SERVER_BASE_PATH + "/todos/" + id);
-      const data = await response.json();
-      setTodo(data);
-    } catch (error) {
-      setTodo(null);
-      console.error(error);
-    }
-  };
-
-  useEffect(() => {
-    fetchTodo();
-  }, []);
+  const { todo } = useFetchTodo();
 
   if (!todo) {
     return <p className="text-center">コンテンツはありません</p>;
